some progress

+++ src/main/java/lukasgol/account/AccountService.java
@@ -3,5 +3,5 @@ package lukasgol.account;
 public interface AccountService {
     Account getAccount(int id);
 
-    void updateBalance(long l);
+    void updateBalance(double l);
 }
