+++ src/main/java/lukasgol/validators/WithdrawalValidator.java
@@ -2,5 +2,5 @@ package lukasgol.validators;
 
 public interface WithdrawalValidator {
 
-    ValidatorResponse validate(int amount, long balance);
+    ValidatorResponse validate(int amount, double balance);
 }
